+++ core/main.ts
@@ -33,16 +33,15 @@ export * from "./src/contrib/entities/objects/Ball";
 export * from "./src/contrib/entities/objects/Goal";
 export * from "./src/contrib/entities/objects/Projectile";
 export * from "./src/contrib/entities/terrain/Background";
-export * from "./src/contrib/entities/terrain/TileFloor";
-export * from "./src/contrib/entities/terrain/TilingSpriteFloor";
+export * from "./src/contrib/entities/terrain/FloorTiles";
 export * from "./src/contrib/entities/terrain/Wall";
 export * from "./src/contrib/entities/ui/Chat";
 export * from "./src/contrib/entities/ui/Cursor";
 export * from "./src/contrib/entities/ui/FpsText";
-export * from "./src/contrib/components/renderables/TextBox";
-export * from "./src/contrib/components/renderables/HealthBar";
-export * from "./src/contrib/components/renderables/DebugBounds";
-export * from "./src/contrib/components/renderables/Button";
+export * from "./src/contrib/renderables/TextBox";
+export * from "./src/contrib/renderables/HealthBar";
+export * from "./src/contrib/renderables/DebugBounds";
+export * from "./src/contrib/renderables/Button";
 export * from "./src/contrib/components/relationships/Controlling";
 export * from "./src/contrib/components/Position";
 export * from "./src/contrib/components/Player";
@@ -59,4 +58,4 @@ export * from "./src/contrib/components/Actions";
 export * from "./src/contrib/components/actions/ZombieMovement";
 export * from "./src/contrib/components/actions/WASDMovement";
 export * from "./src/contrib/components/actions/VehicleMovement";
-export * from "./src/contrib/components/actions/util";
+export * from "./src/contrib/components/actions/PlayerControlsEntity";
