--- core/src/contrib/components/renderables/readme.md
@@ -1,9 +1,0 @@
-## notes on pixi.js graphics
-
-```ts
-// this method resizes the canvas
-renderer.app.renderer.resize(100, 200);
-
-// this callback listens to resize events
-renderer.app.renderer.addListener("resize", () => {});
-```
