+++ core/src/contrib/entities/buttons/DebugButton.ts
@@ -1,13 +1,17 @@
 import { Entity, ValidAction, World } from "@piggo-gg/core";
 import { Button, Clickable, Position } from "@piggo-gg/core";
-import { Text } from "pixi.js";
+import { Graphics, Text } from "pixi.js";
 
 export const DebugButton = (): Entity => {
 
   let pressed = false;
 
+  const outline = new Graphics();
+  const shadow = new Graphics();
+
   const debugButton = Entity({
     id: "debugButton",
+    persists: true,
     components: {
       position: new Position({ x: 40, y: 5, screenFixed: true }),
       clickable: new Clickable({
@@ -15,22 +19,21 @@ export const DebugButton = (): Entity => {
         click: ValidAction((_, world: World) => {
           pressed = !pressed;
           if (pressed) {
-            const r = debugButton.components.renderable as Button;
-            r.shadow.tint = 0x00ffff;
-            r.outline.tint = 0x00ffff;
+            shadow.tint = 0x00ffff;
+            outline.tint = 0x00ffff;
             if (world.renderer) world.debug = true;
           } else {
-            const r = debugButton.components.renderable as Button;
-            r.shadow.tint = 0xffffff;
-            r.outline.tint = 0xffFFFF;
+            shadow.tint = 0xffffff;
+            outline.tint = 0xffFFFF;
             if (world.renderer) world.debug = false;
           }
         })
       }),
-      renderable: new Button({
+      renderable: Button({
         dims: { w: 32, textX: 8, textY: 5 },
         zIndex: 4,
         text: new Text({ text: "🔍", style: { fill: "#FFFFFF", fontSize: 16 } }),
+        outline, shadow
       })
     }
   });
