+++ core/src/contrib/entities/buttons/FullscreenButton.ts
@@ -3,6 +3,7 @@ import { Text } from "pixi.js";
 
 export const FullscreenButton = (id: string = "fullscreenButton") => Entity({
   id: id,
+  persists: true,
   components: {
     position: new Position({
       x: 5, y: 5, screenFixed: true
@@ -19,7 +20,7 @@ export const FullscreenButton = (id: string = "fullscreenButton") => Entity({
         }
       })
     }),
-    renderable: new Button({
+    renderable: Button({
       dims: { w: 32, textX: 8, textY: 5 },
       zIndex: 4,
       text: (new Text({ text: "⚁", style: { fill: "#FFFFFF", fontSize: 16 } }))
