+++ games/strike/Strike.ts
@@ -1,4 +1,4 @@
-import { Background, GameBuilder, Health, HealthBarSystem, PlayerSpawnSystem, Zombie } from "@piggo-gg/core";
+import { Background, GameBuilder, Health, HealthBarSystem, PlayerSpawnSystem, FloorTiles, Zombie } from "@piggo-gg/core";
 
 export const Strike: GameBuilder<"strike"> = ({
   id: "strike",
@@ -8,8 +8,18 @@ export const Strike: GameBuilder<"strike"> = ({
       Zombie({ id: "zombie" }).extend([
         new Health(100, 100)]
       ),
-      Background()
+      Background({img: "stars.png"}),
+      FloorTiles({ rows: 50, cols: 50}),
     ],
     systems: [PlayerSpawnSystem, HealthBarSystem]
   })
 })
+
+/*
+world.addEntity(Projectile({ radius: 5 }));
+
+// Background({img: "stars.png"})
+// Background({img: "aurora.png"})
+// Background({json: { path: "iso-floor-1.json", img: "decor" }})
+// Background({json: { path: "iso-floor-2.json", img: "dark" }})
+*/
