+++ web/res/iso-floor-1.json
@@ -1,0 +1,11 @@
+{
+  "frames": {
+    "decor": { "frame": { "x": 256, "y": 384, "w": 256, "h": 128 } }
+  },
+  "meta": {
+    "image": "iso-floor-1.png",
+    "format": "RGBA8888",
+    "size": { "w": 512, "h": 768 },
+    "scale": "1"
+  }
+}
